ErrorDecoder SessionManager
Implemented ErrorDecoder to capture and handle appropriate responses, ensuring a proper state is returned to our clients. Additionally, implemented logic for session management. When a user logs in, if they remain inactive for 1 hour, we have to decide whether to log them out from the session or reset the session. Currently, the session is reset for inactive users.