Feature create update event (#21)
* Trying to retain localdatetimes values when creating an event , but so far unsuccessfully. Therefore managed to redirect the page and keep the other fields values. Not done yet

* Create Event - Email confirmation

(still thinking of how to retain the values of the LocalDateTime and MultipartFile files..) , the other fields have been taken care of. Also managed to fix the email confirmation logic. Upon confirmation , either success or not , we redirect them to the login page/form.

* Generic servername and port

changed the url to be generic (suitable for local and production environment)

* Generic servername and port - added and passed to the registration end point.