Update event (#23)
* Update-create event

Added redirects after validation exception. In event update request , the event picture is not required anymore , if the user doesn't attach anything , we are not saving image in the db , in general we are not making any requests to the db. NB!! - We have to display a message after successful creation of event.

* Refactor manage events controller

With the new changes in the backend , i had to fit the controller , also added the dto coming from the backend to display all the events of an organisation.